+++ src/command/sample_encode.rs
@@ -94,16 +94,20 @@ pub async fn run(
     let samples = sample_args.sample_count(duration).max(1);
     let temp_dir = sample_args.temp_dir;
 
-    let (samples, full_pass) = {
-        if SAMPLE_SIZE * samples as _ >= Duration::from_secs_f64(duration.as_secs_f64() * 0.85) {
+    let (samples, sample_duration, full_pass) = {
+        if input_is_image {
+            (1, duration.max(Duration::from_secs(1)), true)
+        } else if SAMPLE_SIZE * samples as _
+            >= Duration::from_secs_f64(duration.as_secs_f64() * 0.85)
+        {
             // if the sample time is most of the full input time just encode the whole thing
-            (1, true)
+            (1, duration, true)
         } else {
-            (samples, false)
+            (samples, SAMPLE_SIZE, false)
         }
     };
-
-    bar.set_length(SAMPLE_SIZE_S * samples * 2);
+    let sample_duration_s = sample_duration.as_secs();
+    bar.set_length(sample_duration_s * samples * 2);
 
     // Start creating copy samples async, this is IO bound & not cpu intensive
     let (tx, mut sample_tasks) = tokio::sync::mpsc::unbounded_channel();
@@ -140,7 +144,10 @@ pub async fn run(
             None => break,
         };
         let (sample_idx, sample_n) = (sample_idx as u64, sample_idx + 1);
-        bar.set_prefix(format!("Sample {sample_n}/{samples}"));
+        match full_pass {
+            true => bar.set_prefix("Full pass"),
+            false => bar.set_prefix(format!("Sample {sample_n}/{samples}")),
+        };
 
         let (sample, sample_size) = sample?;
 
@@ -172,7 +179,7 @@ pub async fn run(
         };
         while let Some(progress) = output.next().await {
             if let FfmpegOut::Progress { time, fps, .. } = progress? {
-                bar.set_position(time.as_secs() + sample_idx * SAMPLE_SIZE_S * 2);
+                bar.set_position(time.as_secs() + sample_idx * sample_duration_s * 2);
                 if fps > 0.0 {
                     bar.set_message(format!("enc {fps} fps,"));
                 }
@@ -180,6 +187,7 @@ pub async fn run(
         }
         let encode_time = b.elapsed();
         let encoded_size = fs::metadata(&encoded_sample).await?.len();
+        let encoded_probe = ffprobe::probe(&encoded_sample);
 
         // calculate vmaf
         bar.set_message("vmaf running,");
@@ -187,7 +195,7 @@ pub async fn run(
             &sample,
             args.vfilter.as_deref(),
             &encoded_sample,
-            &vmaf.ffmpeg_lavfi(ffprobe::probe(&encoded_sample).resolution),
+            &vmaf.ffmpeg_lavfi(encoded_probe.resolution),
             enc_args
                 .pixel_format()
                 .max(input_pixel_format.unwrap_or(PixelFormat::Yuv444p10le)),
@@ -201,7 +209,7 @@ pub async fn run(
                 }
                 VmafOut::Progress(FfmpegOut::Progress { time, fps, .. }) => {
                     bar.set_position(
-                        SAMPLE_SIZE_S + time.as_secs() + sample_idx * SAMPLE_SIZE_S * 2,
+                        sample_duration_s + time.as_secs() + sample_idx * sample_duration_s * 2,
                     );
                     if fps > 0.0 {
                         bar.set_message(format!("vmaf {fps} fps,"));
@@ -226,6 +234,7 @@ pub async fn run(
             sample_size,
             encoded_size,
             encode_time,
+            sample_duration: encoded_probe.duration.unwrap_or(sample_duration),
         });
 
         // Early clean. Note: Avoid cleaning copy samples
@@ -236,13 +245,11 @@ pub async fn run(
     }
     bar.finish();
 
-    let input_size = fs::metadata(&*input).await?.len();
-    let predicted_size = results.encoded_percent_size() * input_size as f64 / 100.0;
     let output = Output {
         vmaf: results.mean_vmaf(),
-        predicted_encode_size: predicted_size as _,
-        predicted_encode_percent: results.encoded_percent_size(),
-        predicted_encode_time: results.estimate_encode_time(duration, input_is_image),
+        predicted_encode_size: results.estimate_encode_size(duration, full_pass),
+        encode_percent: results.encoded_percent_size(),
+        predicted_encode_time: results.estimate_encode_time(duration, full_pass),
     };
 
     if !bar.is_hidden() {
@@ -256,8 +263,9 @@ pub async fn run(
         stdout_format.print_result(
             output.vmaf,
             output.predicted_encode_size,
-            output.predicted_encode_percent,
+            output.encode_percent,
             output.predicted_encode_time,
+            input_is_image,
         );
     }
 
@@ -302,12 +310,18 @@ struct EncodeResult {
     encoded_size: u64,
     vmaf_score: f32,
     encode_time: Duration,
+    /// Duration of the sample.
+    ///
+    /// This should be close to `SAMPLE_SIZE` but may deviate due to how samples are cut.
+    sample_duration: Duration,
 }
 
 trait EncodeResults {
     fn encoded_percent_size(&self) -> f64;
     fn mean_vmaf(&self) -> f32;
-    fn estimate_encode_time(&self, input_duration: Duration, image: bool) -> Duration;
+    /// Return estimated encoded **video stream** size.
+    fn estimate_encode_size(&self, input_duration: Duration, single_full_pass: bool) -> u64;
+    fn estimate_encode_time(&self, input_duration: Duration, single_full_pass: bool) -> Duration;
 }
 impl EncodeResults for Vec<EncodeResult> {
     fn encoded_percent_size(&self) -> f64 {
@@ -326,17 +340,31 @@ impl EncodeResults for Vec<EncodeResult> {
         self.iter().map(|r| r.vmaf_score).sum::<f32>() / self.len() as f32
     }
 
-    fn estimate_encode_time(&self, input_duration: Duration, image: bool) -> Duration {
+    fn estimate_encode_size(&self, input_duration: Duration, single_full_pass: bool) -> u64 {
+        if self.is_empty() {
+            return 0;
+        }
+        if single_full_pass {
+            return self[0].encoded_size;
+        }
+
+        let sample_duration: Duration = self.iter().map(|s| s.sample_duration).sum();
+        let sample_factor = input_duration.as_secs_f64() / sample_duration.as_secs_f64();
+        let sample_encode_size: f64 = self.iter().map(|r| r.encoded_size as f64).sum();
+
+        (sample_encode_size * sample_factor).round() as _
+    }
+
+    fn estimate_encode_time(&self, input_duration: Duration, single_full_pass: bool) -> Duration {
         if self.is_empty() {
             return Duration::ZERO;
         }
-        if image {
+        if single_full_pass {
             return self[0].encode_time;
         }
 
-        let sample_factor =
-            input_duration.as_secs_f64() / (SAMPLE_SIZE_S as f64 * self.len() as f64);
-
+        let sample_duration: Duration = self.iter().map(|s| s.sample_duration).sum();
+        let sample_factor = input_duration.as_secs_f64() / sample_duration.as_secs_f64();
         let sample_encode_time: f64 = self.iter().map(|r| r.encode_time.as_secs_f64()).sum();
 
         let estimate = Duration::from_secs_f64(sample_encode_time * sample_factor);
@@ -355,7 +383,7 @@ pub enum StdoutFormat {
 }
 
 impl StdoutFormat {
-    fn print_result(self, vmaf: f32, size: u64, percent: f64, time: Duration) {
+    fn print_result(self, vmaf: f32, size: u64, percent: f64, time: Duration, image: bool) {
         match self {
             Self::Human => {
                 let vmaf = match vmaf {
@@ -375,8 +403,12 @@ impl StdoutFormat {
                     v => style!("{}%", v).bold(),
                 };
                 let time = style(HumanDuration(time)).bold();
+                let enc_description = match image {
+                    true => "image",
+                    false => "video stream",
+                };
                 println!(
-                    "VMAF {vmaf:.2} predicted full encode size {size} ({percent}) taking {time}"
+                    "VMAF {vmaf:.2} predicted {enc_description} size {size} ({percent}) taking {time}"
                 );
             }
             Self::Json => {
@@ -392,10 +424,19 @@ impl StdoutFormat {
     }
 }
 
+/// Sample encode result.
 #[derive(Debug, Clone)]
 pub struct Output {
+    /// Sample mean VMAF score.
     pub vmaf: f32,
+    /// Estimated full encoded **video stream** size.
+    ///
+    /// Encoded sample size multiplied by duration.
     pub predicted_encode_size: u64,
-    pub predicted_encode_percent: f64,
+    /// Sample mean encoded percentage.
+    pub encode_percent: f64,
+    /// Estimated full encode time.
+    ///
+    /// Sample encode time multiplied by duration.
     pub predicted_encode_time: Duration,
 }